use ruleguard in `make lint`
Refs #149
