+++ .github/workflows/go.yml
@@ -8,9 +8,9 @@
       "runs-on": "ubuntu-latest",
       steps: [
         {
-          name: "Set up Go 1.13",
+          name: "Set up Go 1.14",
           uses: "actions/setup-go@v1",
-          "with": {"go-version": 1.13},
+          "with": {"go-version": 1.14},
           id: "go",
         },
         {name: "Check out code into the Go module directory", uses: "actions/checkout@v1"},
