+++ Makefile
@@ -20,6 +20,8 @@ lint:
 	curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(GOPATH_DIR)/bin v1.30.0
 	$(GOPATH_DIR)/bin/golangci-lint run ./analyzer/...
 	$(GOPATH_DIR)/bin/golangci-lint run ./ruleguard/...
+	go build -o go-ruleguard ./cmd/ruleguard
+	./go-ruleguard -rules rules.go ./analyzer/... ./ruleguard/...
 	@echo "everything is OK"
 
 .PHONY: lint test test-master
