+++ go.mod
@@ -4,6 +4,7 @@ go 1.15
 
 require (
 	github.com/google/go-cmp v0.5.2
-	github.com/quasilyte/go-ruleguard/dsl v0.0.0-20201222100711-8bb37f2dbd8a // indirect
+	github.com/quasilyte/go-ruleguard/dsl v0.0.0-20201222100711-8bb37f2dbd8a
+	github.com/quasilyte/go-ruleguard/rules v0.0.0-20201231183845-9e62ed36efe1
 	golang.org/x/tools v0.0.0-20200812195022-5ae4c3c160a0
 )
